Switch from galaxyline to feline
Feline provides a bit more flexibility (as described in it's README) and
also fixes a bug or two that have not been fixed in galaxyline (start
screen being hidden on startup for example). Also it's been about 6
months since galaxyline has seen any activity so it's time to find a new
statusline.